+++ inst/tinytest/test-hypothesis.R
@@ -220,9 +220,9 @@ expect_true("setosa, 0 - setosa, 1" %in% mfx$term)
 # Issue #1092: hypothesis = "mean", "meanother"
 mod <- lm(mpg ~ hp + drat + factor(cyl), data = mtcars)
 p1 <- avg_predictions(mod, by = "cyl")
-p2 <- avg_predictions(mod, by = "cyl", hypothesis = "mean")
+p2 <- avg_predictions(mod, by = "cyl", hypothesis = "meandev")
 expect_equivalent(p2$estimate, p1$estimate - mean(p1$estimate))
-p2 <- avg_predictions(mod, by = "cyl", hypothesis = "meanother")
+p2 <- avg_predictions(mod, by = "cyl", hypothesis = "meanotherdev")
 expect_equivalent(p2$estimate, p1$estimate - (sum(p1$estimate) - p1$estimate) / (nrow(p1) - 1))